requirements.txt
fastapi>=0.121.1
uvicorn>=0.38.0
transformers>=4.57.1
# pinecone-client>=3.0.0,<4.0.0
pinecone>=3.0.0
numpy>=2.2.6
python-dotenv>=1.0.0